refactor: move model conversion to a separate file
